+++ backend/build.gradle
@@ -28,6 +28,7 @@ repositories {
 		implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 		implementation 'org.springframework.boot:spring-boot-starter-validation'
 		implementation 'org.springframework.boot:spring-boot-starter-web'
+		implementation 'org.springframework:spring-webflux'
 		// STOMP
 		implementation 'org.webjars:stomp-websocket:2.3.4'
 		// SockJS
