+++ backend/src/main/java/com/example/backend/webSocket/CustomHandshakeHandler.java
@@ -1,9 +1,8 @@
-package com.example.backend.config;
+package com.example.backend.webSocket;
 
 import org.springframework.http.server.ServerHttpRequest;
 import org.springframework.stereotype.Component;
 import org.springframework.web.socket.WebSocketHandler;
-import org.springframework.web.socket.server.HandshakeHandler;
 import org.springframework.web.socket.server.support.DefaultHandshakeHandler;
 
 import java.security.Principal;
