+++ backend/src/main/java/com/example/backend/webSocket/StompPrincipal.java
@@ -1,4 +1,4 @@
-package com.example.backend.config;
+package com.example.backend.webSocket;
 
 import java.security.Principal;
