+++ backend/src/main/java/com/example/backend/webSocket/WebClientConfig.java
@@ -1,4 +1,4 @@
-package com.example.backend.config;
+package com.example.backend.webSocket;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
